python/requirements.txt
backoff==2.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
Flask==2.2.2
googleapis-common-protos==1.57.1
grpcio @ file:///Users/runner/miniforge3/conda-bld/grpc-split_1670307403148/work
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
opentelemetry-api==1.15.0
opentelemetry-distro==0.36b0
opentelemetry-exporter-otlp==1.15.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-exporter-otlp-proto-http==1.15.0
opentelemetry-instrumentation==0.36b0
opentelemetry-instrumentation-aws-lambda==0.36b0
opentelemetry-instrumentation-dbapi==0.36b0
opentelemetry-instrumentation-flask==0.36b0
opentelemetry-instrumentation-jinja2==0.36b0
opentelemetry-instrumentation-logging==0.36b0
opentelemetry-instrumentation-sqlite3==0.36b0
opentelemetry-instrumentation-urllib==0.36b0
opentelemetry-instrumentation-wsgi==0.36b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
opentelemetry-util-http==0.36b0
protobuf==4.21.12
requests==2.28.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
